news_squeezer/requirement.txt
bert-serving-client==1.9.1
bert-serving-server==1.9.1
pandas>=0.24.2
numpy>=1.16.4
opencv-python==4.1.0.25
keras>=2.2.4
scikit-learn==0.21.2
